Renamed AhSniper package to auctionhouse
