+++ src/main/java/me/partlysanestudios/partlysaneskies/general/WikiArticleOpener.java
@@ -4,7 +4,7 @@ import java.net.URI;
 import java.net.URISyntaxException;
 
 import me.partlysanestudios.partlysaneskies.Main;
-import me.partlysanestudios.partlysaneskies.general.economy.AhSniper.AhGui;
+import me.partlysanestudios.partlysaneskies.general.economy.auctionhouse.AhGui;
 import me.partlysanestudios.partlysaneskies.utils.Utils;
 import net.minecraft.client.gui.inventory.GuiContainer;
 import net.minecraft.inventory.Slot;
